+++ NEWS.md
@@ -18,6 +18,8 @@
 
 * Added a new theme element `tbl_svysummary-arg:missing_stat`
 
+* Fix in `tbl_strata_nested_stack()` when there are unobserved combinations of strata levels. (#2179)
+
 # gtsummary 2.1.0
 
 ### New Features and Functions